feat: overview tab

+++ frontend/lib/utils/money.dart
@@ -1,3 +1,3 @@
-String stringifyCentAmount(int amount) {
+String stringifyCentAmount(double amount) {
   return "${(amount / 100).toStringAsFixed(2)}€";
 }
